roast_my_life/workshop/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.1.1
requests==2.31.0
markdown>=3.4.0